Fix deprecated usage
In response to warning raised by numpy:

DeprecationWarning: `np.float` is a deprecated alias for the builtin
`float`. To silence this warning, use `float` by itself. Doing this will
not modify any behavior and is safe. If you specifically wanted the
numpy scalar type, use `np.float64` here.

+++ test/test_math.py
@@ -4,7 +4,7 @@ import numpy.testing as npt
 import febtools as feb
 
 np.seterr(all="raise")
-_ZERO_ATOL = np.finfo(np.float).resolution * 1000
+_ZERO_ATOL = np.finfo(float).resolution * 1000
 
 
 class TestLinspaced(TestCase):
